refactor(cli): extract sandbox destroy action
