+++ src/lib/nemoclaw-runtime-bridge.ts
@@ -4,7 +4,6 @@
 /* v8 ignore start -- transitional bridge until command actions are extracted from src/nemoclaw.ts. */
 
 export interface NemoClawRuntimeBridge {
-  sandboxDestroy: (sandboxName: string, args?: string[]) => Promise<void>;
   sandboxRebuild: (sandboxName: string, args?: string[]) => Promise<void>;
   upgradeSandboxes: (args?: string[]) => Promise<void>;
 }
